setup: Fix conflicting manifests bug
When trying to add the motorola lux manifest, the script failed, because it
selected bot motorola_lux.xml and sony_pollux.xml, because it used *$DEVICE.xml
to find the manifests. This fixes this for this case, but there's still the
possisbility for a device called xyz_otherdevice_lux.xml to also be selected. To
completely solve this we would need to put all manifests of a vendor into a
seperate directory.