setup: Only sync, don't do breakfast/lunch
This didn't work, because the env variables were not saved in the current shell,
so it was useless. Running the script using `. ./halium/devices/setup` also
doesn't work for breakfast/lunch. If it is possible somehow to do this, though,
of course we can add the build configuration later, again.