setup: Kill the LiveWallPaper feature
Seems that the LiveWallPaper feature in 5.1 causes havoc during build.
It's not an apk/jar so we kill it by removing it completely from the .mk and .sh
files in device, device common, vendor and vendor common folders.

This should help first time porters to get an image to build without forking & manually patching.

Signed-off-by: Herman van Hazendonk github.com@herrie.org