oneplus6: add enchilada and cleanup formatting
The 6 and 6T only differ by a single repo, as we don't have biometrics
support for in screen fingerprint readers, only the enchilada device
will need a system image in order to support both in ubuntu touch (as
far as I can tell). However it makes sense to support both devices here.

Change-Id: I34cc529bd9580706066113fc92cd83f63a428b22